WIP: style-improvement
