+++ src/main/java/com/github/dakusui/actionunit/actions/While.java
@@ -2,7 +2,7 @@ package com.github.dakusui.actionunit.actions;
 
 import com.github.dakusui.actionunit.core.Action;
 import com.github.dakusui.actionunit.core.Context;
-import com.github.dakusui.actionunit.core.generator.ActionGenerator;
+import com.github.dakusui.actionunit.generators.ActionGenerator;
 
 import java.util.function.Predicate;
 import java.util.function.Supplier;
