--- src/main/java/com/github/dakusui/actionunit/core/generator/RunnableGenerator.java
@@ -1,4 +1,0 @@
-package com.github.dakusui.actionunit.core.generator;
-
-public interface RunnableGenerator<I> extends ValueGenerator<I, Runnable> {
-}
