+++ src/main/java/com/github/dakusui/actionunit/generators/ActionGenerator.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.actionunit.core.generator;
+package com.github.dakusui.actionunit.generators;
 
 import com.github.dakusui.actionunit.actions.ValueHolder;
 import com.github.dakusui.actionunit.core.Action;
@@ -6,7 +6,7 @@ import com.github.dakusui.actionunit.core.Context;
 
 import java.util.function.Function;
 
-public interface ActionGenerator<I> extends ValueGenerator<I, Action> {
+public interface ActionGenerator<I> extends Generator<I, Action> {
   static <I> ActionGenerator<I> of(Function<ValueHolder<I>, Function<Context, Action>> func) {
     return valueHolder -> context -> func.apply(valueHolder).apply(context);
   }
