+++ src/main/java/com/github/dakusui/actionunit/generators/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.actionunit.core.generator;
+package com.github.dakusui.actionunit.generators;
 
 public interface BooleanGenerator<I> extends ValueGenerator<I, Boolean> {
 }
