+++ src/main/java/com/github/dakusui/actionunit/generators/StreamGenerator.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.actionunit.core.generator;
+package com.github.dakusui.actionunit.generators;
 
 import java.util.stream.Stream;
