+++ src/main/java/com/github/dakusui/actionunit/generators/ValueGenerator.java
@@ -1,0 +1,6 @@
+package com.github.dakusui.actionunit.generators;
+
+@FunctionalInterface
+public interface ValueGenerator<I, O>
+    extends Generator<I, O> {
+}
