+++ src/test/java/com/github/dakusui/actionunit/examples/RecoveryOperation.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.actionunit.examples;
 
-import com.github.dakusui.actionunit.ActionUnit;
+import com.github.dakusui.actionunit.compat.ActionUnit;
 import com.github.dakusui.actionunit.core.Action;
 import com.github.dakusui.actionunit.core.Context;
 import com.github.dakusui.actionunit.exceptions.ActionException;
@@ -8,7 +8,7 @@ import com.github.dakusui.actionunit.visitors.reporting.ReportingActionPerformer
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static com.github.dakusui.actionunit.core.ActionSupport.*;
+import static com.github.dakusui.actionunit.compat.CompatActionSupport.*;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
 /**
