+++ src/test/java/com/github/dakusui/actionunit/examples/WhileExample.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.actionunit.examples;
 
-import com.github.dakusui.actionunit.ActionUnit;
+import com.github.dakusui.actionunit.compat.ActionUnit;
 import com.github.dakusui.actionunit.core.Action;
 import com.github.dakusui.actionunit.core.Context;
 import com.github.dakusui.actionunit.io.Writer;
