+++ src/test/java/com/github/dakusui/actionunit/ut/BuildersTest.java
@@ -4,8 +4,8 @@ import com.github.dakusui.actionunit.actions.Leaf;
 import com.github.dakusui.actionunit.actions.ValueHolder;
 import com.github.dakusui.actionunit.core.Action;
 import com.github.dakusui.actionunit.core.Context;
-import com.github.dakusui.actionunit.core.ValueHandlerActionFactory;
-import com.github.dakusui.actionunit.core.generator.ActionGenerator;
+import com.github.dakusui.actionunit.compat.ValueHandlerActionFactory;
+import com.github.dakusui.actionunit.generators.ActionGenerator;
 import com.github.dakusui.actionunit.examples.UtContext;
 import com.github.dakusui.actionunit.helpers.Checks;
 import com.github.dakusui.actionunit.io.Writer;
