+++ src/test/java/com/github/dakusui/actionunit/ut/TestMethodNegativeTest.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.actionunit.ut;
 
-import com.github.dakusui.actionunit.ActionUnit;
+import com.github.dakusui.actionunit.compat.ActionUnit;
 import com.github.dakusui.actionunit.core.Action;
 import com.github.dakusui.actionunit.examples.UtContext;
 import org.junit.Test;
