+++ src/test/java/com/github/dakusui/actionunit/ut/actions/CompositeTest.java
@@ -7,7 +7,7 @@ import org.junit.Test;
 import java.util.Collections;
 import java.util.Iterator;
 
-import static com.github.dakusui.actionunit.core.ActionSupport.*;
+import static com.github.dakusui.actionunit.compat.CompatActionSupport.*;
 import static java.util.Collections.singletonList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
